feat(auth): update _generateAuthResponse fn for session

+++ src/modules/session/session.repository.js
@@ -1,35 +1,23 @@
 import { SessionModel } from './session.model.js';
 
 export class SessionRepository {
-    static async createSession(data) {
+    async createSession(data) {
         return await SessionModel.create(data);
     }
-    async findSession({ userId, refreshTokenHash, userAgent }) {
-        return await SessionModel.findOne({
-            user: userId,
-            refreshTokenHash,
-            userAgent,
-            expireAt: { $gt: new Date() },
-        });
-    }
-    async findAllSession({ userId }) {
-        return await SessionModel.findOne({
-            user: userId,
-            expireAt: { $gt: new Date() },
-        });
+
+    async findById(sessionId) {
+        return await SessionModel.findById(sessionId);
     }
 
-    async deleteSession({ userId, refreshTokenHash, userAgent }) {
-        return await SessionModel.deleteOne({
-            user: userId,
-            refreshTokenHash,
-            userAgent,
-        });
+    async findByUserAndAgent(userId, userAgent) {
+        return await SessionModel.findOne({ user: userId, userAgent });
     }
 
-    async deleteAll({ userId }) {
-        return await SessionModel.deleteMany({
-            user: userId,
-        });
+    async deleteByUserAndAgent(userId, userAgent) {
+        return await SessionModel.deleteOne({ user: userId, userAgent });
+    }
+
+    async deleteAll(userId) {
+        return await SessionModel.deleteMany({ user: userId });
     }
 }
